requirements.txt
numpy

Kivy[base]
https://github.com/kivymd/KivyMD/archive/master.zip

pyyaml

pyudev
pyserial

pycodec2
pyaudio

patchelf
nuitka
